requirements.txt
aws-cdk-lib
constructs>=10.0.0,<11.0.0
sagemaker==2.214.3
numpy